Expanded the critical section inside `ProcessManager::spawn`.
This eliminates a race between `process::finalize` and
`ProcessManager::spawn` in which a managed process
(`ProcessManager::spawn(X, true)`) will reference the `gc`
process after `gc` has been destroyed.  The precise
interleaving is possible because the critical section in
`ProcessManager::spawn` only protects access to the
`processes` map.  Meaning that a spawn can pass the check
for `finalizing` and then wait until `gc` has been destroyed.

This commit expands the critical section in `ProcessManager::spawn`
to include the check for `finalizing` as well as the dereferencing
of `gc`.

Review: https://reviews.apache.org/r/57399