Implemented discard behavior in process::Queue.
When a consumer calls `Queue::get` on an empty Queue, the Queue creates
a Promise to back the Future return value.  This Promise currently
does not have a discard handler, which may cause problems if the
Future is chained to multiple continuations.  For example, see
the scenario in MESOS-6919.

This commit implements an (potentially expensive) discard handler
on the Queue's Promise.  If the Future return value is discarded,
the Queue will remove the corresponding Promise from its internal
queue of promises.  The operation is expensive because it needs
to reconstruct the entire internal queue of promises.

Review: https://reviews.apache.org/r/57358