Added logic to check for discard in the server socket's loop.
This deals with a deadlock during libprocess finalization, which
appears as a result of implementing a discard handler on the
`process::Queue` class.  See: https://reviews.apache.org/r/57358/

The libprocess server socket's incoming connections are backed by
a `process::Queue`, which returns Future sockets backed by Promises.
Upon discarding the Future sockets in `process::finalize`, the
discard handlers for the Promises are executed immediately, and in the
same thread.  Because the server socket's `on_accept` loop has an
`.onAny` continuation, this continuation is also executed, leading
to a deadlock on the non-recursive `socket_mutex`.

This commit changes the `.onAny` continuation to stop when it encounters
a discard.  We only discard incoming connections when cleaning up the
socket.

Review: https://reviews.apache.org/r/57821