Changed SSL Socket destruction into a synchronous operation.
Previously, the destruction of SSL Sockets was changed from a relatively
simple synchronous delete of data structures, to one where the data
structures are deleted on the event loop.

This implementation lead to the possibility of the socket's file
descriptor out-living the Socket wrapper.  In the case of server
sockets, this potentially allows incoming connections to be made
against a server socket that is considered "closed" (but the FD is
still accepting).

This commit adds a Gate to the SSL Socket to wait for destruction
to finish before returning.  If the SSL Socket's destructor is itself
run on the event loop, the destructor logic is allowed to short-circuit
and run synchronously.

Review: https://reviews.apache.org/r/58057