Changed libprocess SocketManager to refer to HttpProxy by PID.
HttpProxy actors are spawned to manage incoming HTTP connections.
These actors are themselves garbage collected, meaning it is unsafe
to refer to an HttpProxy by pointer (which is what is currently done).

During libprocess finalization, it is possible for an incoming
connection to spawn an HttpProxy, whose pointer is then deleted by
finalization.  This leads to a potential segfault when cleaning up
the incoming connection, as the SocketManager will dereference
any related HttpProxy actors by pointer.

Review: https://reviews.apache.org/r/58059