CMake: Disabled "local" clusters in the scheduler libraries.
This removes the ability to specify a "local" master in the scheduler
libraries, which subsequently spawns a master and agent.
The consequence of the local cluster is that the scheduler library
is currently required to include the entire master and agent source
tree to instantiate a master/agent in the process.  This dramatically
increases the build cost of this library, but only provides an
oft-unused bit of functionality.

The example framework tests are disabled temporarily because these
tests require the "local" cluster feature.

Review: https://reviews.apache.org/r/69751