CMake: Split part of libmesos into several smaller libraries.
This incrementally splits the gigantic libmesos into some logical
chunks that can be linked to independently, depending on the scope
of the component.

The goal is that all components eventually link to "libmesos-common";
which only contains the 3rdparty libraries, libprocess, the messages
passed between components, and some utilities/helpers; instead of
linking to everything.

This commit also introduces the following two libraries:
  * libmesos-detector: Code used to detect the Mesos master.  Used by
    the Master, Agent, and Scheduler.
  * libmesos-scheduler: Code used by native Schedulers.

Review: https://reviews.apache.org/r/69753