Passed streaming decoder from master /api/v1 to master actor.
This intermediate change passes the incoming streaming handle of a
master SUBSCRIBE request from the Http helpers into the
Master::Subscribers::Subscriber object.

The Subscriber object will be in charge of reading from the streaming
handle and acting on the incoming data.  This logic will be implemented
in a subsequent commit.

Review: https://reviews.apache.org/r/69181