Moved Master::Subscribers::Subscriber constructor into .cpp file.
The logic inside this class's constructor was originally in the master's
header file.  With the introduction of streaming request support (to be
added in subsequent commit(s)), the logic inside the constructor will be
increased; and as such would be better placed outside the header file.

Review: https://reviews.apache.org/r/69182