Added optional heartbeat monitoring to the master /api/v1.
When a client of the master SUBSCRIBE /api/v1 makes a streaming request,
this logic will check that the client sends a heartbeat at least
once every 30 seconds (2x the default heartbeat interval).  If the
heartbeat times out, the connection breaks, or a malformed message is
sent as a heartbeat, the master will actively break the connection.

Review: https://reviews.apache.org/r/69183