Marked internal::recordio::Reader as non-copyable/assignable.
This helper object is basically a wrapper around a libprocess actor.
In the current implementation, if the helper is destructed, the
underlying libprocess actor will be terminated as well.  This means
an errorneous copy of this helper object may result in a dead PID,
which is unguarded and may result in CHECK failures in some scenarios.

Review: https://reviews.apache.org/r/69184