Added streaming header support for /api/v1 SUBSCRIBE.
The existing API allowed the client to send 'Accept' and 'Content-Type'
headers of 'application/json' or 'application/protobuf'.

The SUBSCRIBE endpoint would also take a 'Content-Type' of those two
types, but always return content with 'application/recordio' (because
this is a streaming response).  This maintains that behavior while
also allowing the user to use streaming headers correctly.

Review: https://reviews.apache.org/r/69185