CMake: Reorganized replicated log targets and sources.
Libmesos was originally including 5 sources related to the replicated
log which were not used.  These sources are moved under the mesos-log
binary.

The other replicated log sources are refactored under a new library
which depends on the mesos-detector (primarily for the zookeeper
sources) and leveldb.

The mesos-log binary depends on this new library.

Review: https://reviews.apache.org/r/69766