Added a heartbeat reader to HTTP executor request streams.
This adds a loop, outside of the agent actor, which takes ownership
over reading the heartbeats of a streaming executor SUBSCRIBE request.

This currently has no effect on the executor, as currently, nothing
happens when executors disconnect from the agent; except that messages
are dropped unexpected if this is the case.