CMake: Fixed Java build dependencies.
This hooks up the Java build to pull in the executor, scheduler,
and replicated log libraries.  The Java artifact is then included
as a test dependency.

For OSX, an additional configuration tweak is required to prevent
the CMake FindJNI module from finding an OSX "framework" instead of
a unix-style package (dynamic library).

Review: https://reviews.apache.org/r/69770