CMake: Modified some Linux sources to minimize linkage.
The `mesos-containerizer` helper binary is used by the the agent
and the command executor.  The binary itself requires a limited
amount of sources to build, but on Linux, it pulls in some CNI isolator
code.

This splits the Subcommand used by the CNI isolator into a separate
file that can be shared between the isolator and mesos-containerizer
binary.  Some Linux specific utility methods are gathered into another
library, which will be used by multiple libraries, like the executor,
docker, and agent libraries.