Added test for tearing down frameworks while creating disks.
The CREATE_DISK and DESTROY_DISK operations are "non-speculative"
operations, which means the master must wait for the operations to
complete successfully before the master can update its resources.
Because the master must wait to update the results of non-speculative
operations, it is possible for the framework making the
CREATE/DESTROY_DISK to be torn down before the operation completes.

This commit adds a test to make sure the master can gracefully handle
such a case.

Review: https://reviews.apache.org/r/69869