Accounted for possible lack of OperationID in RP responses.
Operations sent to a resource provider may include an OperationID
in addition to an Operation UUID.  The UUID is sufficient to uniquely
identify the operation, so it is possible for the resource provider
to omit the OperationID in its responses to the agent.

This commit lets the agent fill in any missing ID before updates
are acted upon, such as being sent to the master.

Review: https://reviews.apache.org/r/69872