Handle possible orphaned operations after master/agent failover.
This is one of two possible code paths which can introduce orphaned
operations.

When a master failover occurs, all agents and frameworks must
reregister with the master.  Agents that reregister will report their
operations with an UpdateSlaveMessage.  Any operations without a
known framework will be considered orphans.  Known frameworks are
discovered when the framework reregisters, or an agent running a task
under the framework reregisters.  The race between these reregistrations
will be addressed in a separate commit.

Agent failover can also introduce orphans, if the agent has a pending
operation during failover, and then is migrated to a separate master
before restarting.  This will be handled the same way as agent
reregistration after a master failover.

Review: https://reviews.apache.org/r/69961