SSL Wrapper: Windows: Made sure windows.hpp is included before OpenSSL.
Depending on the version of OpenSSL and what headers of OpenSSL
are included, the order of certain Windows headers may be included
in the "wrong" order.  This results in hundreds of cryptic syntax
errors from MSVC.

This sweeps through the existing sites with OpenSSL headers and
makes sure to include the Windows headers above them.

Review: https://reviews.apache.org/r/71663